Move mocking to setup file
